requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
sqlalchemy
pandas
requests
flask
python-dotenv
